01-linear-regression/requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API Serving
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Configuration
pyyaml>=6.0

# Serialization
joblib>=1.3.0

# Dev & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.0
mypy>=1.5.0
ruff>=0.1.0
black>=23.7.0
